pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident-env"
version = "0.1.0"
description = "On-call incident response RL environment — train AI agents to investigate and resolve production incidents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "incident_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["incident_env", "incident_env.server"]
package-dir = { "incident_env" = ".", "incident_env.server" = "server" }
